Switch `.excluding` to `.where.not` (#10685)
## Context

Excluding will execute the SQL on the query in the argument before
passing the result to the main query causing multiple SQL queries to be
executed. Where not passes the generated SQL to the main query and only
executes 1 SQL statement.

## Changes proposed in this pull request

Switch `.excluding` to `.where.not`

## Guidance to review

- N/A

## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [ ] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency