Refactor FAC filter to throw errors
The FAC filter allows the user to input wrong locations. We need to
throw errors. To do this the filters need to be a form object.

This commit adds this and also a filter component for the find a
candidate view in attempt to not use the generic view component that can
be hard to follow.

Hopefully this would be easier to follow. The component is tested
through the system spec.

The form object has some complexity in it mainly because we need to save
the filters on the provider user and also we need to auto-complete the
location filter.

We need to allow the user to input half completed locations, e.g 'm4'
needs to be autocompleted to 'Manchester M4'.

To do this there's a bit of logic in the initializer that fills the
attributes of the form depending on how the user interacts with the
form.