Make Candidate Pool Invite form use create and update method
Previously, only the create method of the draft_invites_controller was
being hit by the form.

This commit changes the form to use the update method as well, when we
update a draft pool invite.

The form works the same before as now. Because the form object would update or
create based on the presence of id.