Daily summary invite emails (#10712)
## Context

Replacing the single invite email with a combined email. The idea is
that we will send 1 email for all invites over a given time frame. The
initial time frame is set to 30 seconds. This will probably result in a
single email being sent per invite. We will monitor the number of emails
being sent in the first few days of this feature and adjust the timings
as appropriate.

## Guidance to review

Content of mailer to be manually checked
Mailers viewable at 
-
[`/rails/mailers/candidate_mailer/candidate_invites_one_provider`](https://apply-review-10712.test.teacherservices.cloud/rails/mailers/candidate_mailer/candidate_invites_one_provider)
-
[`/rails/mailers/candidate_mailer/candidate_invites_one_provider_two_courses`](https://apply-review-10712.test.teacherservices.cloud/rails/mailers/candidate_mailer/candidate_invites_one_provider_two_courses)
-
[`/rails/mailers/candidate_mailer/candidate_invites_two_providers`](https://apply-review-10712.test.teacherservices.cloud/rails/mailers/candidate_mailer/candidate_invites_two_providers)
-
[`/rails/mailers/candidate_mailer/candidate_invites_multiple_providers`](https://apply-review-10712.test.teacherservices.cloud/rails/mailers/candidate_mailer/candidate_invites_multiple_providers)

![Screenshot 2025-05-15 at 16 11
09](https://github.com/user-attachments/assets/ba7706e1-6f1c-48d4-9f37-ab72bc662a5e)