Require audit comment when bulk email bulk_unsubscribe
Previously when a support user used the bulk_unsubscribe form to stop
candidates from receiving emails. You couldn't add a audit comment.

This was a problem as we could not audit why users where removed from
the emails list or who did it.

This commit fixes this. But sadly we need to iterate over each candidate
and save the audit comment as an individual sql query.

update_all(audit_comment:) doesn't work. But it's in the same
transaction at least.