Refactor Pool::Candidates class
This class now has been refactored to return all the applications forms
that can be in the pool by a live query. This is the `application_forms_in_the_pool`
method

.application_forms_for_provider also has changed, it doesn't need to
accept providers. We don't have provider dismissals at this point.
We also don't need to check if the candidate opted in or their account
is blocked.

The CandidatePoolApplication does that for us.

I've also put the curated_application_forms back as a private method.

And the filtered_application_forms is scoped to CandidatePoolApplication
tables