Scope the candidate preferences to the filtered candidates
Previously, we would search in all candidate location preferences
before serving the results for find a candidate.

This meant that we would call `.near` on more records that we needed.

This commit limits the distance calculations the geocoder needs to do to
only the candidates that are filtered, only their candidate location
preferences. Not all candidate location preferences