Fix race condition on pool invite form
There is an edge case when 2 provider users can invite the same
candidate to apply for a course in a short amount of time.

This causes an issue because the pool invite status changes from draft
to published. The second provider user would get an error because we
can't find the invite.

This commit fixes this by redirecting the second provider user to the
candidate show page, where they can start inviting the candidate again,
If they choose the same course they will see an error.

We need to redirect the user because they are trying to update a
pool_invite that is published. We need to stop this, so we redirect the
user back to the candidate show page