Changed the subject filter attribute on the FAC page
We need to change the subject filter name, from subject to subject_ids.
To allow subjects to be viewed in requests. There's a redacted function
that is applied to the subject name in logs. To redact the subject in
email.

In doing this we also need an easier way to be able to change filter
attributes on the FAC page. The issue comes from saving the filter names
in DB, on the ProviderUser model.

If an attribute changes or is removed the filters would break for the
people who have the old filter attributes saved in DB.

This commit makes the change process easier. To change an attribute we
need to add an alias.

If we forget to add an alias, the form will still work, because we
remove any filter attributes that the CandidatePoolFilter class doesn't
support.

But if we want to persist the filters even after we change the names, we
need to add an alias.

Users will be automatically switched to the new filter name as they use
the service, their filters in db will match the new filters, because we
save the filters on every request. So in theory after some time the
alias can be removed.