Make the candidate_pool_filter more robust
Previously, if a filter that is not in the candidate_pool_filter
attributes list is in db, on the ProviderUser. The filter would blow up.

This will check if the filters in the DB can be assigned. This will
remove the 'bad filters' saved in DB, allowing us to remove filters from
candidate_pool_filter without worrying what filters are saved in DB.