Remove one login accounts on sandbox
Our sandbox env is using the production env of one login. This was
because the testing integration env one login has required an extra step
to login. There was a pop-up that required you to input a separate
account and password from your one login.

This is not the case anymore and we have created quite a few fraud
alerts in one login because of the sandbox accounts created

We will switch sandbox to use the testing intgration env of one login,
similar to what QA uses. Part of doing that we need to wipe the one
login accounts on sandbox.

This commit adds a data migration to do that. We will run it manually.