Allow provider to add a message to their invites
We need to allow a provider to add their custom message to the
invitation email the candidate gets from FAC.

This commit adds a new step to the invite form which allow the user to
add a custom message on the Pool::Invite record which will be used in
the email template of the invite.

The message will work with individual mail deliveries or bulk ones.

The new form step also needs to show a preview of how the email would
look like. The email template in the preview is hard-coded.

There's no easy way to reuse the email template we send to candidates.