Fix pool_invite form to use the update method
Previously the update method in the draft_invites controller wasn't
actually used.

This didn't really cause an issue as the create method can handle
updates as well, because of the form object save method which works with
updates and creates.

Now that the form needs to grow and we need to add more steps, we need
to fix this.

This commit just adds the correct paths and form methods to the new and
edit partials of the draft_invite form.