Redirect to FAC index from show page if the application is not found
Candidates go in an out of the pool whilst a provider browses the FAC
list. If they don't refresh the page and go to the show page of a
candidate that is no longer in the pool they can encounter an error.

This commit fixes this, redirecting the provider to the index page if
the candidate is no longer in the pool