Add New, Viewed and invited to FAC index page
This commit adds these 3 statuses and the functionality to mark
candidates as viewed.

Initially all candidates would be New unless they've already been
invited.

Once the provider user lands on the show page of a candidate that
candidate will be marked as viewed. Through the ProviderPoolAction
table.

This is specific to a provide user.

The invited status is org wide. Meaning any provider user that invites a
candidate marks that candidate as invited.