Migrate provider user fac filters to provider_user_filters table
This will be ran manually to minimise the gap between switching to the
new filters structure.

There will still be a gap where a provider user
might set a new filter but I think this would not be a problem. Worst
case scenario they would lose that filter but once they apply it back
they will always have that filter set for them.