Preload the degree qualifications
Previously we would have made 20 queries to DB to show the degree
qualifications for each candidate on the FAC index page.

This commit fixes this by preloading the degrees, effectively using 1 query
to fetch all the degrees for all 20 candidates

Similar with the candidates.

We need to preload degrees because it messes up the site_distance
virtual attribute if we use includes on degrees. It seems to go missing.