New status tab for FAC
We need a new tab on the Find a candidate page. This tab is only for not
seen candidates.

We need to exclude the candidates that have been viewed and invited by

We also need to persist filters from the all tab to the new tab and
vice-versa

This is done in the form objects, similar to callbacks. They are not on
the model because I wanted to keep this logic for the form object in
case things change in the future.

We relay on the most up to date filter between the all and new tab to
know which tab we should redirect the user to after they submit an invite.

So if you land on the new tab. The filters that are the most up to date
are the new tab filters. That's how we know where to redirect the user
after invite. So that we don't pass params around.