Redirect to invited tab after invite if the user came from invited tab
Now that the Find a candidate view has multiple tabs we should redirect
the user to the correct tab after invite. We do this by saving the
filters in the ProviderUserFilter on every tab load. This way we can get
the last updated_at filter record and redirect the user to the correct
tab.

Based on the idea that their latest tab is the tab they came from. Not
100% robust but good enough.