Add copy link to find a candidate show page
We want to make it easier for provider users to share candidates from
find a candidate.

This adds a share button on the show page that redirects to a page where
you can copy the show link of a FAC candidate. This is meant to be
shared with someone else.

If javascript is disabled the copy to clipboard button will not show.