Update course recommender to check for courses (#10875)
## Context

Before sending a Candidate an email with the Recommendation link, we
want to check that the Find service has got courses to recommend based
on the parameters set.

## Changes proposed in this pull request

Add a check to get the number of courses that would be recommended, as
part of the link generation

## Guidance to review

- NA

## Concerns

- Calling the UI is necessary, however this should only ever be done in
a background job. The current sample links are rendered in the Support
Interface, these should be removed.


## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [ ] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency