Candidate funding type preference migrations
We want to allow candidates to express their preference to what funding
type courses they want to be invited.

This adds migrations for that but also on the candidate_pool_applications
table so that the providers can filter by this preference.

The column on the candidate_pool_applications is a boolean, the filter
in the provider interface is going to only filter for candidates that
are interested in fee funded courses.