Candidate funding type preference
We want to allow the candidate to express their preference to what
course funding type they want to be invited to.

We want to add another step in the candidate preferences that asks the
candidate to express this preference.

This new step is going to be shown conditionally. If the candidate has
applied to any fee course we will not show this new step and assume they
are interested in fee courses.

We only show this new step if the candidate only applied to salary
courses.

Because we are introducing this step now, when candidates have
set their preferences already. We will give the candidates that applied
only to salary courses to specific this preference when they go back and
edit their preferences.

Work on the provider side to use this funding type preference will
follow.