Allow providers to filter by fee based courses in FAC
We allowed candidates to express their preference to be invited to fee
funded courses or not.

We need to allow providers to filter by this preference.

We will allow providers just to exclude candidates that are not
interested in fee based courses through their preferences or through
their choices.

Candidates that don't have a preference but only applied to salary
courses will be excluded by this filter.

This filter just uses a boolean course_funding_type_fee on the
candidate_pool_applications table. This is set by the FAC worker every
15 minutes