Add application_choice_id and candidate_decision to invite
In a lot of cases we want to get the application_choice that links to a
pool_invite, we often do this through the course and course option.

It would be easier if we save an application_choice_id on the invite at
the time of submitting an application.

This just adds the migration, later PRs will follow to use this column

It also adds a candidate_decision enum, with 'not_responded' by default.

The value of this column will change with a later PR that will backfill
application_choices