Link invite to application choice on course change
There is an edge case where a course can change on an application choice
for an application form that has an invite to the new course on the
application_choice.

In this case we want to link this application_choice to the invite if
the courses match. The new course with the course on the invite.

Doing it in update_course_option_and_associated_fields method ensures
that we do this link every time a course has changed, including the API.