Backfill applied application_choice on invites
We are linking application choices to invites when candidates submit
their applications.

We now need to backfill this. To link the application choices to the
invites if the courses match.

The query in the data migration is based on scopes we already have but
we need to join on application choices.