Wipe the immigration status when support user changes right to work
When a support user changes the right to work of a candidate the
immigration status is not changed.

This commit will set the immigration status to nil if the right to work
is not true.

There's also an edge case where the support user needs to set the right
to work back to true after setting it to false. In this case we
redirect the support user to fill in an immigration status otherwise the
candidate will not be able to access their own personal details