Add course_closed to invite
We need to show the candidate if the course of their invite has closed
or not.

We can store a boolean on the invite table so we don't have to join with
the course on every invite the index page.

This boolean will be kept in sync by the sync with publish job.