Fix course status column on pool invite
This removes the course_closed column from the pool invite and adds a
course_open column that is meant to do the same thing.

But this is easier to read when we'll have to keep this in sync.

We can do invite.update(course_open: course.open?) rather than negate
course.open?