Link the invite to an application choice from creation of draft
The way we used to link the application_choice to an invite was flawed.

When a candidate created a draft application choice for a invited
course. We would set the `candidate_decision` to `applied` but the
application_choice wasn't saved to DB yet, so the choice id was not saved on
the invite.

This caused some errors and some data inconstancies where invites had
the `candidate_decision` set to `applied` but no choice id.

This commit fixes this by always linking the choice to the invite if the
candidate applied to invited course. We link at the time of draft choice
creation. Doing this will allow us to not try to link again when
candidate submits the choice.

This is OK with product as it's still showing the intent of the
candidate to apply to invite.

There's also an edge case when a candidate creates a draft application
choice for invited course and changes to another course they have not
been invited. In this case we remove the link between the choice and the
invite. We know we can do this because the application choice original
course and current course are the same. Even if the candidate changes
courses of a draft choice.

If the choice has the original course same as an invited course we don't
remove the link between invite and application choice.