Hide Referee input until Reference has been submitted (#11028)
## Context

It was possible for Providers to see information from Referees, before
the Reference was fully submitted. Information such as feedback would be
visible before the Referee clicked "Submit reference"

## Changes proposed in this pull request

- Wrap and apply logic for showing Referee input in the view component
- Update the rendering of the component to handle collections (not
related but a nice tidy up)

## Guidance to review

No change required on the API as this check is already in place.

### Before change

1. As a Candidate in the Offer Accepted state (pending conditions) and
with incomplete References
2. As a Referee, complete the Reference up to the “Check your reference
for Candidate Name” page - do not “Submit reference”
3. As a Provider, view the Application’s References
4. See the feedback from the unsubmitted Reference

### After change

1. As a Candidate in the Offer Accepted state (pending conditions) and
with incomplete References
2. As a Referee, complete the Reference up to the “Check your reference
for Candidate Name” page - do not “Submit reference”
3. As a Provider, view the Application’s References
4. Unsubmitted Reference feedback is not visible

This change applies to Confidential, Feedback, Safeguarding and
Corrections to Relationship rows