Allow provider users to add a description to api tokens (#11023)
## Context

This PR builds on the basic functionality of allowing provider users to
add a description to the tokens they create -- this will make the more
easily identifiable if they wish to revoke them.

All of this is behind a feature flag. 

The next PR in this series will be to add some ui for updating the
`manage_api_tokens` permission -- it only exists as a column in the db
that has to be updated on the rails console at the moment. Then we'll do
some more thorough testing and make sure provider users are aware of it.

## Changes proposed in this pull request


https://github.com/user-attachments/assets/0d8a1367-6c41-40b1-8c49-1080e703ba28

## Guidance to review
- Locally, turn on the api_token_manage feature flag
- For simplicity, make all the users able to add tokens
(`ProviderPermissions.update_all(manage_api_tokens: true)`)
- Login as any provider and add a token. You should be required to add a
description, and you should see the description on the table when you
view it.
- Login as support and view the api tokens page. You should see any
descriptions you've added.

---------

Co-authored-by: David Young <david@dcyoung.dev>