Fix flakey test
Unrelated to this PR.

Test was flaky due to checking for the text "Change" not being present. Occasionally the degree name included the work "change" e.g "BA Change Management".
We now explicitly look for the change link. Oddly this is a support interface link that is rendered in the shared DegreeComponent