Use accepted pool_invite#candidate_decision
This commit changes all the places we use `pool_invite.applied` to
`pool_invite.accepted`

The enum is present. A data migration will run when this change is
deployed. To migrate all the pool_invite.applied to accepted

If there was a brief period of limbo, the code should not break. As the
enum is present just no records have it set.

The views should not break. There might be some content inconsistencies.
Where the `applied` status would not be shown.

Once the data migration has run it will work as expected.