Allow candidates to respond to invites that had a draft before the invite
There's an edge case where candidates cannot respond to an invite if
they already had a draft application choice for the invited course.

This commit fixes this by linking the choice to the invite when they
'accept' the invite. This only happens if we find an existing draft
application choice.

If we don't find one we just ask them to create a draft application
choice, which will link the choice to the invite, when the draft is
created.

I've also moved all the state management of the candidate_decision into
the CandidateInterface::InviteApplication to keep everything in one
place.