Add open date for FAC
We don't want to open the FAC feature when apply opens. We want to open
it on the 20th of November.

This commit adds some logic to always open the FAC on the 20th of
November, after apply opens. Regardless of the cycle year. It will
always stay close between cycles and between apply opening and the 20th
of November.