Change cycle dates generator
Apply and find along with policy have agreed to change the cycle dates.

The main change concerns find opening. Find will open on the closest
Tuesday in working days to the 1st of October. And based on that other
dates changed, like when apply opens.

This commit changes the generator to apply this rule