Rework and remove the start-carry-over endpoint (#11092)
## Context

A bug was reported that the Application list filter tabs not working
during end of cycle deadlines.

To reduce complexity we want to remove the `start-carry-over` page
moving the content and functionality into the `choices` page.

As a prerequisite, we will combine the logic to display the correct
content into a single view component.

## Changes proposed in this pull request

This PR will start to consolidate the `start-carry-over` endpoint into
the `choices` endpoint.

The bug reported is due to the Candidate being redirected to the
`start-carry-over` page which is designed to look like the `choices`
page.

This PR consolidates the Carry Over messaging into the Application
Choices index. The Carry Over messaging will never show under the
current setup as the User will still be redirected if they are able to
Carry Over.
Following PRs will remove this redirect allowing the content to be
displayed appropriately.

## Guidance to review

- No visible change to the UI
- Review commit by commit