Update get_into_teaching gem
We need to use the one login gem so that we can implement the
backchannel functionality.

The issue with the get_into_teaching gem is that it uses faraday 1.x.
And one_login was using 2.13. This was causing a conflict that was not
easily fixed.

A separate branch from master is open on the get_into_teaching gem that uses
faraday 2.13. This plays nicely with the one_login faraday.

https://github.com/DFE-Digital/get-into-teaching-api-ruby-client/pull/106

This branch will not be merged into master yet. The team that maintains
this gem wants to test it in their own time. So this means we are the
guinea pigs here.

Don't think this will cause any issues though. Faraday should work as
before.