Refactors Application States into Data classes
This allows easier visibility over which states are included/exclude in each defined constant.
This also allows for easy testing of Application States.

This next step is to remove the external use of the constants in application_state_change.rb in favour of the scope methods.

The application_state_change_spec.rb is currently double testing that the scopes match the constants and also the exact values returned by each. This will enable the removal of the constants, while ensuring nothing changes