Implement one login backchannel
We want to Implement the backchannel feature in apply. So that we logout
the user if they logout somewhere else using their one login account.

To be able to do this easily, we need to use the one login gem.

This commit adds this gem and changes the omniauth config to work with
the gem. And adds a backchannel_logout endpoint where one login will
post requests to us.

We will use the gem to decode the sub and find the user with it. Delete
the session which will log them out.