Update service information banner (#11148)
We want to upgrade our DB version, first on sandbox then production. We
expect about 3 hours of donwtime.

So we need to inform our users of this downtime. This commit updates the
service banner that's behind a feature flag now.

We can turn the feature flag on after merging this.