This update allows developers to declare partial controller paths
This is needed as some flows have multiple controllers, eg deferred_offer has `check`, `course`, `study_mode` `site`, `conditions`.
The current setup would require each of these controller names to be added to the list.
With this change it allows us to declare a partial path to the controller, ie `provider_interface/deferred_offer` which is the start of the controller path for each of the deferred offer controllers.