Allow active navigation links to be declared by partial controller paths (#11163)
## Context

This is needed as some flows have multiple controllers, eg
deferred_offer has `check`, `course`, `study_mode` `site`, `conditions`.
The current setup would require each of these controller names to be
added to the list.

## Changes proposed in this pull request

With this change it allows us to declare a partial path to the
controller, ie `provider_interface/deferred_offer` which is the start of
the controller path for each of the deferred offer controllers.