Updated the API scopes to view the current and next cycles (#11216)
## Context

The GIT CRM integration is failing on some candidates who have skipped a
recruitment cycle year - they were last active in 2024 cycle, dormant
through 2025, and have now carried over into 2026. The current scopes
only looked at the current cycle or the candidate was created in since
the end of the last cycle 🙃

## Changes proposed in this pull request

These changes to the scopes are applied to v1.2 as this is the version
of the API that GIT CRM team are using. More recent versions, 1.3 and
1.4 have not been used in the past month.

## Guidance to review

- N/A


## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [ ] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency