Previous teacher training form (#11177)
This commit adds the new previous_teacher_training form.
We save the yes\no answer from the user as it's easier for review pages
to work as the user needs to be able to go back to see the details
they've entered.

We also need to update the
application_form.previous_teacher_training_completed which marks the
section as completed.

We create a draft from a published previous_teacher_training if the user
is editing a published previous_teacher_training. When they publish the
draft, we delete any other published previous_teacher_trainings.