Make the invite status on pool invites more robust
There is a bug in the way we set the candidate_decision when a candidate
declined an invite and then creates an application choice for the invited course.

This sets the candidate_decision to `accepted`. But when they delete
their that application choice or they change the course. We need to set
the candidate_decision back to `declined` if they declined it in the
first place.

We know they declined it because we still have a declined reason
attached to the invite, even if the candidate decides to create an
application choice for the invited course, triggering a change in
candidate_decision to `accepted`

Previously we would set the candidate_decision back to `not_responded`.

This was a problem as the candidate already responded to the invite, by
declining. So they will not get any chasers or banners prompting them to
respond to the invite.