[lb/1300] Update application sharing information page (#11232)
## Context

Improvements to the FAC content and flow after user research.

## Changes proposed in this pull request

- Updated content for 'how application sharing works'
- Updated flow -- the user is taken directly to the Opt in / out
question from submitting an application rather than to the 'how
application sharing works page.
- Deleted the ShareDetails component -- none of the functionality was
relevant anymore, so just moved the text to the index page where it
lived.

## Guidance to review

The review app is set up so that we are mid cycle 2026. 
- Add an application, you should be directed to the preferences flow
without the interruption page
- Click on 'Application sharing' and then 'Read about how sharing your
application details works' You should see the new content.
- The back link will take you to the main Application sharing page. 


https://github.com/user-attachments/assets/e46441a8-87a7-4bb9-8393-62e9c6a4507d

Note, I've not use locale files because the entire page is text, there
isn't any functionality, or conditional logic. so I think it is easier
to read this way. I don't think we have a clear strategy for translation
files, but let me know if you disagree with this approach.

## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency