Confirm Deferrals flow (#11197)
## Context

We want to guide the User through the process of Confirming a Deferred
offer. In the case where the original course is still available in the
current cycle, the User is taken directly to the Check page.

The User can choose to change any part of the deferral. They may only
select Study Modes and Locations that are associated to the Course. If
they choose to change the Course and either of the previously selected
Study Mode or Location is not available, the User will be directed to
choose one.

## Changes proposed in this pull request

- Enabled the flow for the confirm deferral process
- Tidied up the UI and select methods
- Refactored controllers to share functionality
- Added a root path for Confirm Deferrals

## Guidance to review

**Review app**
Deferred Application where course is not available -
https://apply-review-11197.test.teacherservices.cloud/provider/applications/4/deferred-offer/check

Deferred Application with available course -
https://apply-review-11197.test.teacherservices.cloud/provider/applications/5/deferred-offer/check



https://github.com/user-attachments/assets/9ebedfc1-a522-42d7-a272-d9a3da28408e