Rubocop config - stop using frozen string literal magic comment at the top of files (#11243)
## Context

We no longer need the `frozen_string_literal: true` magic comment at the
top of each file. The need for it went away with ruby 3.4, and we are on
ruby 3.4.4.

## Changes proposed in this pull request

Rubocop configuration, and then run with the rule set to never require
the magic comment.

## Guidance to review

Any concerns? Objections?

## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [ ] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency