[lb1121] More restrictive checks on quickfire sign up (#11246)
## Context

We were too successful with the quickfire sign up (interruption page for
candidates once they have completed certain sections of the application
form) and had to deactivate the feature flag so that the TTA team could
meet the demand.

We have agreed more restrictive checks for showing people the
interruption page.

## Changes proposed in this pull request

in order to see the interruption page, the candidate now has to have:
- a bachelor degree from the UK (before could have been master,
doctorate, or bachelor from uk or abroad)
- Must be from a subject in high demand. 

The high demand subjects are tricky. We provided the team with a list of
all the degrees entered from the past year and they selected the ones
which should be included. The ones that account for about 3/4 (12907 of
16893) of the matches are directly mapped to subjects chosen from the
drop down list of subjects (from the DFE reference data), either from
the degree_subject_uuid (the large majority) or the subject synonyms /
matching data in the reference gem.

That leaves us with 3986 free text, most of which have only been entered
once, so are unlikely to be entered again. They are full of spelling
mistakes, sometimes specific grades, qualification, and descriptions,
this probably won't catch many more candidates in practice.

I played around with some fuzzy matching formulas, but not sure it is
really worth the effort and could result in more matches than we
actually want.

## Guidance to review

[This is a
candidate](https://apply-review-11246.test.teacherservices.cloud/support/applications/71)
who currently sees the interruptions because they meet all the criteria,
including a
- UK bachelor degree, 
- desireable subject and a 
- grade of 2:2 or higher.

Sign in as them and make changes to the degree. As long as you meet the
criteria above, you should see the interruption page. (just don't ever
dismiss or sign up for an adviser, you only see it until you respond).

## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency

---------

Co-authored-by: Catalin Voineag <11318084+CatalinVoineag@users.noreply.github.com>